Document refresh interval and refactor clerk api functions
- Move clerk api functions to clerk.ts.
- Add comment calrifying the 25s token refresh.
