+++ frontend/src/components/ClerkSessionRefresher.tsx
@@ -1,7 +1,9 @@
 import { useAuth } from '@clerk/clerk-react';
 import { useEffect } from 'react';
-import { refreshClerkSession } from '../lib/api';
+import { refreshClerkSession } from '../lib/clerk';
 
+// Clerk session tokens only live ~60s as part of Clerk's hybrid auth model (https://clerk.com/docs/guides/how-clerk-works/overview).
+// Refreshing just under halfway (25s) keeps our backend's registered token comfortably ahead of expiry so server requests stay authenticated.
 const CLERK_SESSION_REFRESH_INTERVAL_MS = 25_000;
 
 export function ClerkSessionRefresher(): null {
@@ -26,4 +28,3 @@ export function ClerkSessionRefresher(): null {
 
   return null;
 }
-
